Tests use resource managed file buffer fixture
